[net.bioclipse.sdk.templating] new Templater.java
This class was developed with TDD in less than an hour. It does string
templating, using a template string and a set of key/value pairs for
replacement. The templater replaces occurrences of the substring
'${somekey}' in the template string. There can be several occurrences of
the same key in the template string, but passing the same key twice, or
not passing a key found in the template string, results in an exception.